+++ index.html
@@ -19,6 +19,6 @@
   </div>
 </body>
 
-<script src="script.js"></script>
+<script type="module" src="script.js"></script>
 
 </html>
