+++ script.js
@@ -1,4 +1,8 @@
 "use strict";
+
+import { getComments, sendComment } from "./api.js";
+import { commentsInnerHTML, formInnerHTML } from "./render.js";
+
 let nameInputEl = document.getElementById('name-input');
 let textInputEl = document.getElementById('text-input');
 let textErrorEl = document.getElementById('text-error');
@@ -7,7 +11,6 @@ let sendButtonEl = document.getElementById('send-button');
 const commentsEl = document.getElementById('comments');
 const addFormEl = document.getElementById('form');
 const deleteButtonEl = document.getElementById('delete-button');
-const optionsDate = ["ru-RU", { day: "2-digit", month: "2-digit", year: "2-digit", hour: "2-digit", minute: "2-digit" }];
 
 let isLoadComment = false;
 let comments = [];
@@ -80,120 +83,8 @@ function disableButton(button, ...fields) {
 // Удаление последнего комментария
 deleteButtonEl.addEventListener("click", () => {
     comments.pop();
-    console.log(comments);
-    renderComments();
+    // renderComments();
 });
-
-// Получает данные с API, преобразует в нужный формат
-const getComments = () => {
-    return fetch("https://wedev-api.sky.pro/api/v1/anton-pashinov/comments", { method: "GET" })
-        .then((response) => {
-            return response.json();
-        })
-        .then((responseData) => {
-            comments = responseData.comments.map((item) => {
-                item["name"] = item["author"]["name"];
-                item["date"] = new Date(Date.parse(item["date"])).toLocaleDateString(...optionsDate);
-                item["is_liked"] = item["isLiked"];
-                item["editing"] = false;
-                item["isLikeLoading"] = false;
-                delete item["author"];
-                delete item["isLiked"];
-
-                return item;
-            });
-        });
-};
-
-// Рендер формы и добавление обработчиков для ее элементов
-const renderForm = () => {
-    if (isLoadComment) {
-        addFormEl.innerHTML = `<img src="loader.gif" alt="загрузка">`;
-    } else {
-        addFormEl.innerHTML = `<div class="add-form">
-  <div class="add-form-error add-form-error__hide" id="name-error"></div>
-  <input type="text" class="add-form-name" placeholder="Введите ваше имя" id="name-input" value="${userNameInput}"/>
-  <div class="add-form-error add-form-error__hide" id="text-error"></div>
-  <textarea type="textarea" class="add-form-text" placeholder="Введите ваш коментарий" rows="4"
-  id="text-input">${userTextInput}</textarea>
-  <div class="add-form-row">
-    <button disabled class="add-form-button" id="send-button">Написать</button>
-  </div>
-</div>`;
-        nameInputEl = document.getElementById('name-input');
-        textInputEl = document.getElementById('text-input');
-        textErrorEl = document.getElementById('text-error');
-        nameErrorEl = document.getElementById('name-error');
-        sendButtonEl = document.getElementById('send-button');
-
-        fields = [
-            { "input": nameInputEl, "err": nameErrorEl, "message": "Заполните имя" },
-            { "input": textInputEl, "err": textErrorEl, "message": "Напишите текст" }
-        ];
-
-        nameInputEl.addEventListener('input', () => disableButton(sendButtonEl, ...fields));
-        textInputEl.addEventListener('input', () => disableButton(sendButtonEl, ...fields));
-
-        textInputEl.addEventListener('keyup', (event) => {
-            if (event.keyCode === 13 && event.ctrlKey) {
-                sendButtonEl.click();
-            }
-        });
-
-        // Обработка события Click для кнопки отправки нового комментария
-        sendButtonEl.addEventListener('click', () => {
-            if (!formValidate(...fields)) return;
-
-            isLoadComment = true;
-            renderForm();
-
-            userNameInput = nameInputEl.value
-                .replaceAll("&", "&amp;")
-                .replaceAll("<", "&lt;")
-                .replaceAll(">", "&gt;");
-            userTextInput = textScreen(textInputEl.value);
-
-            fetch("https://wedev-api.sky.pro/api/v1/anton-pashinov/comments",
-                {
-                    method: "POST",
-                    body: JSON.stringify({
-                        text: userTextInput,
-                        name: userNameInput,
-                        forceError: true,
-                    }),
-                })
-                .then((response) => {
-                    if (response.status === 400) {
-                        throw new Error("Текст должен быть не короче трех символов");
-                    };
-                    if (response.status === 500) {
-                        throw new Error("Ошибка сервера");
-                    };
-                })
-                .then(() => {
-                    return getComments();
-                })
-                .then(() => {
-                    return renderComments();
-                })
-                .then(() => {
-                    isLoadComment = false;
-                    userNameInput = "";
-                    userTextInput = "";
-                    renderForm();
-                })
-                .catch((error) => {
-                    if (error.message === "Ошибка сервера") {
-                        sendButtonEl.click();
-                    } else {
-                        alert(error.message);
-                        isLoadComment = false;
-                        renderForm();
-                    }
-                })
-        });
-    };
-};
 
 // Кнопки лайк
 const likeEventListeners = () => {
@@ -291,49 +182,8 @@ const commentEventListeners = () => {
     };
 };
 
-// Рендерит комментарии
 const renderComments = () => {
-    const commentsHtml = comments
-        .map((comment) => {
-            let activeLike = comment.is_liked ? " -active-like" : "";
-            comment.isLikeLoading ? activeLike += " -loading-like" : activeLike;
-
-            let commentBody = comment.text;
-            let button = `<button class="edit-button" data-id="${comment.id}">Редактировать</button>`;
-
-            if (comment["editing"]) {
-                const commentTextScreen = commentBody
-                    .replaceAll(`<div class="quote">`, `[quote]`)
-                    .replaceAll(`</div>`, `[/quote]`);
-                commentBody =
-                    `<div class="add-form-error add-form-error__hide" id="text-error-${comment.id}"></div>
-            <textarea type="textarea" class="add-form-text edit-field" placeholder="Введите ваш коментарий" rows="4"
-          id="text-input-${comment.id}">${commentTextScreen}</textarea>`;
-                button = `<button class="edit-button save-button" data-id="${comment.id}">Сохранить</button>`;
-            };
-
-            return `<li class="comment" data-id="${comment.id}">
-            <div class="comment-header">
-              <div>${comment.name}</div>
-              <div>${comment.date}</div>
-            </div>
-            <div class="comment-body">
-              <div class="comment-text">
-                ${commentBody}
-              </div>       
-            </div>
-            <div class="comment-footer">
-              <div class="likes">
-                <span class="likes-counter">${comment.likes}</span>
-                <button class="like-button${activeLike}" data-id="${comment.id}"></button>
-              </div>
-            </div>
-            ${button}
-          </li>`
-        })
-        .join("");
-    commentsEl.innerHTML = commentsHtml;
-
+    commentsInnerHTML(comments, commentsEl);    
     likeEventListeners();
     editFieldEventListener();
     editBtnEventListeners();
@@ -341,8 +191,74 @@ const renderComments = () => {
 }
 
 
+// Рендер формы и добавление обработчиков для ее элементов
+const renderForm = () => {
+    formInnerHTML(isLoadComment, addFormEl, userNameInput, userTextInput)
+    if (!isLoadComment) {
+        nameInputEl = document.getElementById('name-input');
+        textInputEl = document.getElementById('text-input');
+        textErrorEl = document.getElementById('text-error');
+        nameErrorEl = document.getElementById('name-error');
+        sendButtonEl = document.getElementById('send-button');
+
+        fields = [
+            { "input": nameInputEl, "err": nameErrorEl, "message": "Заполните имя" },
+            { "input": textInputEl, "err": textErrorEl, "message": "Напишите текст" }
+        ];
+
+        nameInputEl.addEventListener('input', () => disableButton(sendButtonEl, ...fields));
+        textInputEl.addEventListener('input', () => disableButton(sendButtonEl, ...fields));
+
+        textInputEl.addEventListener('keyup', (event) => {
+            if (event.keyCode === 13 && event.ctrlKey) {
+                sendButtonEl.click();
+            }
+        });
+
+        // Обработка события Click для кнопки отправки нового комментария
+        sendButtonEl.addEventListener('click', () => {
+            if (!formValidate(...fields)) return;
+
+            isLoadComment = true;
+            renderForm();
+
+            userNameInput = nameInputEl.value
+                .replaceAll("&", "&amp;")
+                .replaceAll("<", "&lt;")
+                .replaceAll(">", "&gt;");
+            userTextInput = textScreen(textInputEl.value);
+
+            sendComment(userNameInput, userTextInput)
+                .then(() => {
+                    return getComments();
+                })
+                .then((data) => {
+                    comments = data;
+                    return renderComments();
+                })
+                .then(() => {
+                    isLoadComment = false;
+                    userNameInput = "";
+                    userTextInput = "";
+                    renderForm();
+                })
+                .catch((error) => {
+                    if (error.message === "Ошибка сервера") {
+                        sendButtonEl.click();
+                    } else {
+                        alert(error.message);
+                        isLoadComment = false;
+                        renderForm();
+                    }
+                })
+        });
+    };
+};
+
+
 getComments()
-    .then(() => {
+    .then((data) => {
+        comments = data;
         renderComments();
     })
     .then(() => {
