+++ styles.css
@@ -54,6 +54,7 @@ body {
   display: flex;
   flex-direction: column;
   white-space: pre-line;
+  overflow-wrap: break-word;
 }
 .comment-footer {
   display: flex;
